Rename common port package to common
