+++ LabApiUtilities/src/main/com/microsoft/identity/labapi/utilities/authentication/common/CertificateCredential.java
@@ -20,7 +20,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.labapi.utilities.authentication.port;
+package com.microsoft.identity.labapi.utilities.authentication.common;
 
 import java.io.IOException;
 import java.security.KeyStore;
