+++ LabApiUtilities/src/main/com/microsoft/identity/labapi/utilities/authentication/common/ClientAssertion.java
@@ -20,19 +20,21 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.labapi.utilities.authentication.port;
+package com.microsoft.identity.labapi.utilities.authentication.common;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.experimental.Accessors;
 
 /**
- * Configuration information for the client certificate to be used.
+ * A class representing a client assertion used by the authorization server to authenticate the client application.
+ * Adding support for client authentication for internal use with test execution
+ * https://tools.ietf.org/html/rfc7521#section-6.1
  */
 @Getter
 @Accessors(prefix = "m")
 @AllArgsConstructor
-public class ClientCertificateMetadata {
-    private final String mAlias;
-    private final char[] mPassword;
+public class ClientAssertion {
+    private final String mClientAssertion;
+    private final String mClientAssertionType;
 }
